Moving board_deparse to boards.py
